--- create_database.py
@@ -1,34 +1,0 @@
-import sqlite3
-
-
-# DB CREATION
-conn = sqlite3.connect("people.db")
-columns = [
-    "id INTEGER PRIMARY KEY",
-    "lname VARCHAR UNIQUE",
-    "fname VARCHAR",
-    "timestamp DATETIME",
-]
-create_table_cmd = f"CREATE TABLE person ({','.join(columns)})"
-conn.execute(create_table_cmd)
-
-# DATA ADDITION
-conn = sqlite3.connect("people.db")
-people = [
-    "1, 'Fairy', 'Tooth', '2022-10-08 09:15:10'",
-    "2, 'Ruprecht', 'Knecht', '2022-10-08 09:15:13'",
-    "3, 'Bunny', 'Easter', '2022-10-08 09:15:27'",
-]
-for person_data in people:
-    insert_cmd = f"INSERT INTO person VALUES ({person_data})"
-    conn.execute(insert_cmd)
-conn.commit()
-
-# DATA DISPLAY
-conn = sqlite3.connect("people.db")
-cur = conn.cursor()
-cur.execute("SELECT * FROM person")
-
-people = cur.fetchall()
-for person in people:
-    print(person)
