+++ src/__main__.py
@@ -1,8 +1,8 @@
 import os
 from flask import render_template
 
-import config
-from models import Person
+from src import config
+from src.models import Person
 
 
 app = config.connex_app
@@ -12,6 +12,7 @@ app.add_api(os.path.join(config.basedir, "swagger.yml"))
 @app.route("/")
 def home():
     people = Person.query.all()
+    print(os.listdir('.'))
     return render_template("home.html", people=people)
 
 if __name__ == "__main__":
