+++ src/people.py
@@ -1,8 +1,8 @@
 from typing import *
 from flask import abort, make_response
 
-from config import db
-from models import Person, people_schema, person_schema
+from src.config import db
+from src.models import Person, people_schema, person_schema
 
 
 def create(person: Dict[str,str]) -> Tuple[Dict[str,str],int]:
